text-buffer: widen chunk width past u16
Lines longer than 65535 columns saturated chunk width at u16
max, so cursor, viewport, and rope weight drifted. Store width
as u32 so horizontal extent stays accurate for very long lines.

+++ packages/core/src/zig/text-buffer-segment.zig
@@ -37,7 +37,7 @@ pub const TextChunk = struct {
     mem_id: u8,
     byte_start: u32,
     byte_end: u32,
-    width: u16,
+    width: u32,
     flags: u8 = 0,
     graphemes: ?[]GraphemeInfo = null,
     wrap_offsets: ?[]utf8.WrapBreak = null,
